refactor: split backport-audit into modular package
Convert single-file script into Python package to enable parallel
development of remaining features.

Structure:
- backport_audit/__init__.py - Package init with VERSION
- backport_audit/models.py - Exceptions and dataclasses
- backport_audit/slack.py - Slack user mapping
- backport_audit/config.py - Config and arg parsing
- backport_audit/github_client.py - GitHubClient
- backport_audit/jira_client.py - JiraClient
- backport_audit/utils.py - Utility functions
- backport-audit.py - Entry point (imports from package)

This enables Tasks 8-9 to run in parallel (different files).
