+++ scripts/ci/backport-audit.py
@@ -1,430 +1,18 @@
 #!/usr/bin/env python3
 """
-Audit backport PRs and Jira issues for consistency and completeness.
+Backport PR audit tool - entry point.
+
+This script validates backport PRs and Jira issues for release management.
 """
 
-import argparse
-import base64
-import json
-import os
-import re
-import subprocess
 import sys
-import traceback
-from dataclasses import dataclass
-from datetime import datetime
-from typing import Any, Dict, List, Optional
-from urllib.request import Request, urlopen
-from urllib.error import HTTPError, URLError
 
-VERSION = "1.0.0"
-
-# Slack user ID mapping (GitHub login -> Slack member ID)
-# Source: scripts/ci/get-slack-user-id.sh
-SLACK_USER_MAP: Dict[str, str] = {
-    '0x656b694d': 'U02MJ72K1B5',
-    'BradLugo': 'U042Z3TSZU3',
-    'GrimmiMeloni': 'U048VH2JZ1C',
-    'JoukoVirtanen': 'U033Y28GYN4',
-    'Maddosaurus': 'U01Q5L5R0GJ',
-    'Molter73': 'U02A292NPV2',
-    'RTann': 'U01NZ6U730X',
-    'SimonBaeumer': 'U01Q5RMEHCK',
-    'Stringy': 'U02KJKREKPY',
-    'ajheflin': 'U087GT2H45Q',
-    'akameric': 'U076CG62KL4',
-    'AlexVulaj': 'U03M3QKBES2',
-    'alanonthegit': 'U01PZFFSZRB',
-    'alwayshooin': 'U01PLAWUU8N',
-    'bradr5': 'U03UQ9DM44U',
-    'c-du': 'U02NE59PHT3',
-    'charmik-redhat': 'U035YKHMXEW',
-    'clickboo': 'U01PFFU0YKD',
-    'dashrews78': 'U03FB5XE10V',
-    'daynewlee': 'U03J855QWHF',
-    'dcaravel': 'U04DF45CXBJ',
-    'dvail': 'U032WL9RM53',
-    'ebensh': 'U01Q7HTJ126',
-    'erthalion': 'U02SV8VE3K3',
-    'gaurav-nelson': 'U01P6PMFGKF',
-    'guzalv': 'U08NQKQJH4N',
-    'house-d': 'U03H69TFKH9',
-    'janisz': 'U0218FUVDMJ',
-    'johannes94': 'U03E2SD2ZPB',
-    'jschnath': 'U03AA9E6B09',
-    'jvdm': 'U02TTV416HY',
-    'kovayur': 'U033ZSBGEUQ',
-    'ksurabhi91': 'U043ZP4RN76',
-    'kurlov': 'U035001CQCV',
-    'kylape': 'UGJML86DD',
-    'ludydoo': 'U04TFDR57KQ',
-    'lvalerom': 'U02SJTV567N',
-    'mclasmeier': 'U02DKH1LQ5N',
-    'mfosterrox': 'U01PMH71ACU',
-    'msugakov': 'U020QJZCQAH',
-    'mtodor': 'U039LQ48PT7',
-    'ovalenti': 'U03F2F9EXUL',
-    'parametalol': 'U02MJ72K1B5',
-    'pedrottimark': 'U01RN8V8DEH',
-    'porridge': 'U020XCUG2LA',
-    'rhybrillou': 'U02GPRG4NHF',
-    'robbycochran': 'U03NAEPKDE1',
-    'rukletsov': 'U01G6P17RTK',
-    'sachaudh': 'U01QLCGS0NM',
-    'stehessel': 'U02SDMERUFP',
-    'sthadka': 'U029PASTL5C',
-    'tommartensen': 'U040F2EG19U',
-    'vikin91': 'U02L405V2GH',
-    'vjwilson': 'U01PKQQF0KY',
-    'vladbologa': 'U03NFNXKPH9',
-    'vulerh': 'U02A9CAR59T',
-}
-
-
-def get_slack_mention(github_login: str) -> str:
-    """
-    Get Slack mention for GitHub user.
-
-    Args:
-        github_login: GitHub username
-
-    Returns:
-        Slack mention string (<@ID>, @username, or :konflux:)
-    """
-    if github_login == 'app/red-hat-konflux':
-        return ':konflux:'
-
-    slack_id = SLACK_USER_MAP.get(github_login)
-    if slack_id:
-        return f'<@{slack_id}>'
-    return f'@{github_login}'
-
-
-# Exception hierarchy
-class BackportAuditError(Exception):
-    """Base exception for backport audit tool."""
-
-
-class GitHubError(BackportAuditError):
-    """GitHub API/CLI error."""
-
-
-class JiraError(BackportAuditError):
-    """Jira API error."""
-
-
-@dataclass
-class PR:
-    """Pull request data."""
-    number: int
-    title: str
-    author: str
-    base_ref: str
-    jira_keys: List[str]
-    body: str
-
-
-@dataclass
-class JiraIssue:
-    """Jira issue data."""
-    key: str
-    summary: str
-    fix_versions: List[str]
-    affected_versions: List[str]
-    assignee: Optional[str]
-    team: Optional[str]
-    component: Optional[str]
-
-
-@dataclass
-class ReleaseBranch:
-    """Release branch with version info."""
-    name: str
-    expected_version: str
-    latest_tag: Optional[str]
-
-
-@dataclass
-class Config:
-    """Configuration from environment and arguments."""
-    jira_user: str
-    jira_token: str
-    jira_base_url: str = "redhat.atlassian.net"
-    jira_project: str = "ROX"
-    github_token: Optional[str] = None
-    output_dir: str = "."
-    report_file: str = "backport-audit-report.md"
-    slack_payload_file: str = "slack-payload.json"
-    branches: str = "all"
-    github_run_url: Optional[str] = None
-
-    @classmethod
-    def from_env(cls, args: argparse.Namespace) -> 'Config':
-        """Create config from environment and CLI args."""
-        jira_user = os.environ.get('JIRA_USER')
-        jira_token = os.environ.get('JIRA_TOKEN')
-
-        if not jira_user:
-            raise BackportAuditError("JIRA_USER environment variable is required")
-        if not jira_token:
-            raise BackportAuditError("JIRA_TOKEN environment variable is required")
-
-        return cls(
-            jira_user=jira_user,
-            jira_token=jira_token,
-            jira_base_url=os.getenv('JIRA_BASE_URL', 'redhat.atlassian.net'),
-            jira_project=os.getenv('JIRA_PROJECT', 'ROX'),
-            github_token=os.getenv('GITHUB_TOKEN'),
-            output_dir=args.output_dir,
-            branches=args.branches,
-            github_run_url=args.github_run_url,
-        )
-
-
-class GitHubClient:
-    """GitHub operations via gh CLI."""
-
-    def fetch_prs(self, label: str = "backport", state: str = "open") -> List[Dict[str, Any]]:
-        """
-        Fetch PRs using gh CLI.
-
-        Args:
-            label: Label to filter by
-            state: PR state (open, closed, all)
-
-        Returns:
-            List of PR dictionaries
-        """
-        cmd = [
-            'gh', 'pr', 'list',
-            '--repo', 'stackrox/stackrox',
-            '--search', f'label:{label} draft:false',
-            '--state', state,
-            '--limit', '1000',
-            '--json', 'number,title,author,baseRefName,body,state'
-        ]
-
-        try:
-            result = subprocess.run(
-                cmd,
-                capture_output=True,
-                text=True,
-                check=True,
-                timeout=60
-            )
-            return json.loads(result.stdout)
-        except subprocess.CalledProcessError as e:
-            raise GitHubError(f"Failed to fetch PRs: {e.stderr}")
-        except subprocess.TimeoutExpired:
-            raise GitHubError("gh CLI command timed out")
-        except json.JSONDecodeError as e:
-            raise GitHubError(f"Invalid JSON from gh CLI: {e}")
-
-    def get_pr_details(self, pr_number: int) -> Dict[str, Any]:
-        """
-        Get PR details via gh CLI.
-
-        Args:
-            pr_number: PR number
-
-        Returns:
-            PR details dictionary
-        """
-        cmd = ['gh', 'pr', 'view', str(pr_number), '--json', 'author,body']
-
-        try:
-            result = subprocess.run(
-                cmd,
-                capture_output=True,
-                text=True,
-                check=True,
-                timeout=30
-            )
-            return json.loads(result.stdout)
-        except subprocess.CalledProcessError as e:
-            raise GitHubError(f"Failed to fetch PR #{pr_number}: {e.stderr}")
-        except subprocess.TimeoutExpired:
-            raise GitHubError(f"gh CLI command timed out for PR #{pr_number}")
-        except json.JSONDecodeError as e:
-            raise GitHubError(f"Invalid JSON from gh CLI for PR #{pr_number}: {e}")
-
-    def get_issue_events(self, pr_number: int) -> List[Dict[str, Any]]:
-        """
-        Get issue events via gh API.
-
-        Args:
-            pr_number: PR number
-
-        Returns:
-            List of event dictionaries
-        """
-        cmd = [
-            'gh', 'api',
-            f'repos/stackrox/stackrox/issues/{pr_number}/events'
-        ]
-
-        try:
-            result = subprocess.run(
-                cmd,
-                capture_output=True,
-                text=True,
-                check=True,
-                timeout=30
-            )
-            return json.loads(result.stdout)
-        except subprocess.CalledProcessError as e:
-            raise GitHubError(f"Failed to fetch events for PR #{pr_number}: {e.stderr}")
-        except subprocess.TimeoutExpired:
-            raise GitHubError(f"gh API command timed out for PR #{pr_number}")
-        except json.JSONDecodeError as e:
-            raise GitHubError(f"Invalid JSON from gh API for PR #{pr_number}: {e}")
-
-
-class JiraClient:
-    """Jira REST API client using urllib."""
-
-    def __init__(self, user: str, token: str, base_url: str = "redhat.atlassian.net"):
-        self.user = user
-        self.token = token
-        self.base_url = base_url
-        self._auth_header = self._make_auth_header()
-
-    def _make_auth_header(self) -> str:
-        """Create Basic Auth header."""
-        credentials = f"{self.user}:{self.token}"
-        encoded = base64.b64encode(credentials.encode()).decode()
-        return f"Basic {encoded}"
-
-    def get_issue(self, issue_key: str) -> Optional[JiraIssue]:
-        """
-        Fetch Jira issue via REST API.
-
-        Args:
-            issue_key: Jira issue key (e.g., ROX-12345)
-
-        Returns:
-            JiraIssue or None if not found
-        """
-        fields = "fixVersions,versions,summary,status,assignee,components,customfield_10001"
-        url = f"https://{self.base_url}/rest/api/3/issue/{issue_key}?fields={fields}"
-
-        req = Request(url)
-        req.add_header("Authorization", self._auth_header)
-        req.add_header("Content-Type", "application/json")
-
-        try:
-            with urlopen(req, timeout=30) as response:
-                data = json.loads(response.read())
-                return self._parse_issue(data)
-        except HTTPError as e:
-            if e.code == 404:
-                print(f"WARNING: Jira issue {issue_key} not found", file=sys.stderr)
-                return None
-            print(f"WARNING: HTTP error fetching {issue_key}: {e}", file=sys.stderr)
-            return None
-        except URLError as e:
-            print(f"WARNING: Network error fetching {issue_key}: {e}", file=sys.stderr)
-            return None
-        except json.JSONDecodeError as e:
-            print(f"WARNING: Invalid JSON from Jira for {issue_key}: {e}", file=sys.stderr)
-            return None
-
-    def _parse_issue(self, data: Dict[str, Any]) -> JiraIssue:
-        """Parse Jira API response into JiraIssue."""
-        fields = data.get('fields', {})
-
-        fix_versions = [v['name'] for v in fields.get('fixVersions', [])]
-        affected_versions = [v['name'] for v in fields.get('versions', [])]
-
-        assignee = None
-        if fields.get('assignee'):
-            assignee = fields['assignee'].get('displayName')
-
-        team = None
-        if fields.get('customfield_10001'):
-            team = fields['customfield_10001'].get('name')
-
-        components = [c['name'] for c in fields.get('components', [])]
-        component = ', '.join(components) if components else None
-
-        return JiraIssue(
-            key=data['key'],
-            summary=fields.get('summary', ''),
-            fix_versions=fix_versions,
-            affected_versions=affected_versions,
-            assignee=assignee,
-            team=team,
-            component=component
-        )
-
-    def search_issues(self, jql: str, max_results: int = 1000) -> List[JiraIssue]:
-        """
-        Search Jira issues via JQL.
-
-        Args:
-            jql: JQL query string
-            max_results: Maximum results to return
-
-        Returns:
-            List of JiraIssue objects
-        """
-        from urllib.parse import urlencode
-
-        params = urlencode({
-            'jql': jql,
-            'fields': 'key,summary',
-            'maxResults': max_results
-        })
-        url = f"https://{self.base_url}/rest/api/3/search?{params}"
-
-        req = Request(url)
-        req.add_header("Authorization", self._auth_header)
-        req.add_header("Content-Type", "application/json")
-
-        try:
-            with urlopen(req, timeout=30) as response:
-                data = json.loads(response.read())
-                issues = []
-                for issue_data in data.get('issues', []):
-                    issues.append(JiraIssue(
-                        key=issue_data['key'],
-                        summary=issue_data['fields'].get('summary', ''),
-                        fix_versions=[],
-                        affected_versions=[],
-                        assignee=None,
-                        team=None,
-                        component=None
-                    ))
-                return issues
-        except (HTTPError, URLError, json.JSONDecodeError) as e:
-            print(f"WARNING: Error searching Jira: {e}", file=sys.stderr)
-            return []
-
-
-def parse_args() -> argparse.Namespace:
-    """Parse command-line arguments."""
-    parser = argparse.ArgumentParser(
-        description='Audit backport PRs and validate Jira issues for release management.'
-    )
-    parser.add_argument(
-        '--branches',
-        default='all',
-        help='Release branches (comma-separated or "all")'
-    )
-    parser.add_argument(
-        '--output-dir',
-        default='.',
-        help='Output directory for reports'
-    )
-    parser.add_argument(
-        '--github-run-url',
-        help='GitHub Actions run URL for Slack link'
-    )
-    return parser.parse_args()
+from backport_audit import __version__
 
 
 def main():
-    print(f"Backport Audit Tool v{VERSION}")
+    """Main entry point."""
+    print(f"Backport Audit Tool v{__version__}")
     print("Script structure complete. Ready for implementation.")
     sys.exit(0)
