+++ scripts/ci/backport_audit/jira_client.py
@@ -1,0 +1,130 @@
+"""Jira client using urllib."""
+
+import base64
+import json
+import sys
+from typing import Any, Dict, List, Optional
+from urllib.error import HTTPError, URLError
+from urllib.parse import urlencode
+from urllib.request import Request, urlopen
+
+from .models import JiraIssue
+
+
+class JiraClient:
+    """Jira REST API client using urllib."""
+
+    def __init__(self, user: str, token: str, base_url: str = "redhat.atlassian.net"):
+        self.user = user
+        self.token = token
+        self.base_url = base_url
+        self._auth_header = self._make_auth_header()
+
+    def _make_auth_header(self) -> str:
+        """Create Basic Auth header."""
+        credentials = f"{self.user}:{self.token}"
+        encoded = base64.b64encode(credentials.encode()).decode()
+        return f"Basic {encoded}"
+
+    def get_issue(self, issue_key: str) -> Optional[JiraIssue]:
+        """
+        Fetch Jira issue via REST API.
+
+        Args:
+            issue_key: Jira issue key (e.g., ROX-12345)
+
+        Returns:
+            JiraIssue or None if not found
+        """
+        fields = "fixVersions,versions,summary,status,assignee,components,customfield_10001"
+        url = f"https://{self.base_url}/rest/api/3/issue/{issue_key}?fields={fields}"
+
+        req = Request(url)
+        req.add_header("Authorization", self._auth_header)
+        req.add_header("Content-Type", "application/json")
+
+        try:
+            with urlopen(req, timeout=30) as response:
+                data = json.loads(response.read())
+                return self._parse_issue(data)
+        except HTTPError as e:
+            if e.code == 404:
+                print(f"WARNING: Jira issue {issue_key} not found", file=sys.stderr)
+                return None
+            print(f"WARNING: HTTP error fetching {issue_key}: {e}", file=sys.stderr)
+            return None
+        except URLError as e:
+            print(f"WARNING: Network error fetching {issue_key}: {e}", file=sys.stderr)
+            return None
+        except json.JSONDecodeError as e:
+            print(f"WARNING: Invalid JSON from Jira for {issue_key}: {e}", file=sys.stderr)
+            return None
+
+    def _parse_issue(self, data: Dict[str, Any]) -> JiraIssue:
+        """Parse Jira API response into JiraIssue."""
+        fields = data.get('fields', {})
+
+        fix_versions = [v['name'] for v in fields.get('fixVersions', [])]
+        affected_versions = [v['name'] for v in fields.get('versions', [])]
+
+        assignee = None
+        if fields.get('assignee'):
+            assignee = fields['assignee'].get('displayName')
+
+        team = None
+        if fields.get('customfield_10001'):
+            team = fields['customfield_10001'].get('name')
+
+        components = [c['name'] for c in fields.get('components', [])]
+        component = ', '.join(components) if components else None
+
+        return JiraIssue(
+            key=data['key'],
+            summary=fields.get('summary', ''),
+            fix_versions=fix_versions,
+            affected_versions=affected_versions,
+            assignee=assignee,
+            team=team,
+            component=component
+        )
+
+    def search_issues(self, jql: str, max_results: int = 1000) -> List[JiraIssue]:
+        """
+        Search Jira issues via JQL.
+
+        Args:
+            jql: JQL query string
+            max_results: Maximum results to return
+
+        Returns:
+            List of JiraIssue objects
+        """
+        params = urlencode({
+            'jql': jql,
+            'fields': 'key,summary',
+            'maxResults': max_results
+        })
+        url = f"https://{self.base_url}/rest/api/3/search?{params}"
+
+        req = Request(url)
+        req.add_header("Authorization", self._auth_header)
+        req.add_header("Content-Type", "application/json")
+
+        try:
+            with urlopen(req, timeout=30) as response:
+                data = json.loads(response.read())
+                issues = []
+                for issue_data in data.get('issues', []):
+                    issues.append(JiraIssue(
+                        key=issue_data['key'],
+                        summary=issue_data['fields'].get('summary', ''),
+                        fix_versions=[],
+                        affected_versions=[],
+                        assignee=None,
+                        team=None,
+                        component=None
+                    ))
+                return issues
+        except (HTTPError, URLError, json.JSONDecodeError) as e:
+            print(f"WARNING: Error searching Jira: {e}", file=sys.stderr)
+            return []
